+++ src/main/java/dev/radworks/radiation/RadiationSourceType.java
@@ -7,7 +7,10 @@ public enum RadiationSourceType {
     BLOCK_ITEM_HANDLER("block_item_handler"),
     BLOCK_FLUID_HANDLER("block_fluid_handler"),
     CREATE_TRANSIENT_ITEM("create_transient_item"),
-    CREATE_TRANSIENT_FLUID("create_transient_fluid");
+    CREATE_TRANSIENT_FLUID("create_transient_fluid"),
+    ENTITY_DROPPED_ITEM("entity_dropped_item"),
+    ENTITY_ITEM_FRAME("entity_item_frame"),
+    ENTITY_PLAYER_INVENTORY_AURA("entity_player_inventory_aura");
 
     private final String id;